sidebar refactored
